Move runner and launch into their own files
